app/requirements.txt
django==4.1.7
djangorestframework==3.14.0
django-filter==22.1
psycopg2-binary==2.9.5
asgiref==3.6.0
gunicorn==20.1.0
pytz==2022.7.1
sqlparse==0.4.3
